moved SourceDrawer component to correct path
